vector-store: move index module to vs_index
Rename the index/ directory to vs_index/ to make room for a parallel
full-text-search index module. This commit only moves the files and
updates the module path so git records the rename and preserves history;
type names are left unchanged and refactored in a follow-up commit.
