+++ crates/vector-store/src/engine.rs
@@ -15,8 +15,8 @@ use crate::db::DbExt;
 use crate::db_index::DbIndex;
 use crate::db_index::DbIndexExt;
 use crate::factory::IndexFactory;
-use crate::index::Index;
-use crate::index::factory::IndexConfiguration;
+use crate::vs_index::Index;
+use crate::vs_index::factory::IndexConfiguration;
 use crate::indexes::IndexEntry;
 use crate::indexes::Indexes;
 use crate::memory;
