+++ crates/vector-store/src/httproutes.rs
@@ -12,8 +12,8 @@ use crate::SimilarityScore;
 use crate::distance;
 use crate::engine::Engine;
 use crate::engine::EngineExt;
-use crate::index::IndexExt;
-use crate::index::validator;
+use crate::vs_index::IndexExt;
+use crate::vs_index::validator;
 use crate::indexes;
 use crate::indexes::Indexes;
 use crate::info::Info;
