+++ crates/vector-store/src/monitor_items.rs
@@ -8,8 +8,8 @@ use crate::DbIndexedRow;
 use crate::DbIndexedValue;
 use crate::IndexKey;
 use crate::Metrics;
-use crate::index::Index;
-use crate::index::IndexExt;
+use crate::vs_index::Index;
+use crate::vs_index::IndexExt;
 use crate::perf;
 use crate::table::Operation;
 use crate::table::TableAdd;
