+++ crates/vector-store/src/vs_index/actor.rs
