+++ crates/vector-store/src/vs_index/mod.rs
