+++ crates/vector-store/src/vs_index/opensearch.rs
@@ -14,9 +14,9 @@ use crate::Limit;
 use crate::PartitionId;
 use crate::SpaceType;
 use crate::Vector;
-use crate::index::actor::Index;
-use crate::index::factory::IndexConfiguration;
-use crate::index::validator;
+use crate::vs_index::actor::Index;
+use crate::vs_index::factory::IndexConfiguration;
+use crate::vs_index::validator;
 use crate::memory::Memory;
 use crate::perf;
 use crate::table::IndexIdGenerator;
