+++ crates/vector-store/src/vs_index/usearch.rs
@@ -13,10 +13,10 @@ use crate::Limit;
 use crate::Quantization;
 use crate::SpaceType;
 use crate::Vector;
-use crate::index::actor::AnnR;
-use crate::index::actor::Index;
-use crate::index::factory::IndexConfiguration;
-use crate::index::validator;
+use crate::vs_index::actor::AnnR;
+use crate::vs_index::actor::Index;
+use crate::vs_index::factory::IndexConfiguration;
+use crate::vs_index::validator;
 use crate::memory::Allocate;
 use crate::memory::Memory;
 use crate::memory::MemoryExt;
@@ -1201,7 +1201,7 @@ mod tests {
     use super::*;
     use crate::Config;
     use crate::IndexKey;
-    use crate::index::IndexExt;
+    use crate::vs_index::IndexExt;
     use crate::memory;
     use crate::table::IndexIdGenerator;
     use crate::table::MockTableSearch;
