+++ crates/vector-store/src/vs_index/validator.rs
